bugfix: move function getListener() to pkg/net/listener.go
