+++ cri/v1alpha1/cri_stream.go
@@ -6,12 +6,12 @@ import (
 
 	"github.com/alibaba/pouch/cri/stream"
 	"github.com/alibaba/pouch/daemon/mgr"
-	pouchnet "github.com/alibaba/pouch/pkg/net"
+	"github.com/alibaba/pouch/pkg/netutils"
 )
 
 func newStreamServer(ctrMgr mgr.ContainerMgr, address string, port string) (Server, error) {
 	if address == "" {
-		a, err := pouchnet.ChooseBindAddress(nil)
+		a, err := netutils.ChooseBindAddress(nil)
 		if err != nil {
 			return nil, fmt.Errorf("failed to get stream server address: %v", err)
 		}
