+++ cri/v1alpha2/service/cri.go
@@ -1,13 +1,10 @@
 package service
 
 import (
-	"net"
-	"os"
-	"syscall"
-
 	runtime "github.com/alibaba/pouch/cri/apis/v1alpha2"
 	cri "github.com/alibaba/pouch/cri/v1alpha2"
 	"github.com/alibaba/pouch/daemon/config"
+	"github.com/alibaba/pouch/pkg/netutils"
 
 	"google.golang.org/grpc"
 )
@@ -27,8 +24,6 @@ func NewService(cfg *config.Config, criMgr cri.CriMgr) (*Service, error) {
 		criMgr: criMgr,
 	}
 
-	// TODO: Prepare streaming server.
-
 	runtime.RegisterRuntimeServiceServer(s.server, s.criMgr)
 	runtime.RegisterImageServiceServer(s.server, s.criMgr)
 
@@ -37,12 +32,7 @@ func NewService(cfg *config.Config, criMgr cri.CriMgr) (*Service, error) {
 
 // Serve starts grpc server.
 func (s *Service) Serve() error {
-	// Unlink to cleanup the previous socket file.
-	if err := syscall.Unlink(s.config.CriConfig.Listen); err != nil && !os.IsNotExist(err) {
-		return err
-	}
-
-	l, err := net.Listen("unix", s.config.CriConfig.Listen)
+	l, err := netutils.GetListener(s.config.CriConfig.Listen, nil)
 	if err != nil {
 		return err
 	}
