+++ pkg/netutils/doc.go
@@ -1,4 +1,4 @@
-package net
+package netutils
 
 // NOTE: The code in this package is directly copy from package "k8s.io/apimachinery/pkg/util/net".
 // We do this because we don't want to vendor too many irrelevant packages and try to make
