+++ pkg/netutils/interface.go
@@ -1,4 +1,4 @@
-package net
+package netutils
 
 import (
 	"bufio"
