+++ pkg/netutils/interface_test.go
@@ -1,4 +1,4 @@
-package net
+package netutils
 
 import (
 	"fmt"
