+++ test/z_cli_daemon_test.go
@@ -389,14 +389,14 @@ func (suite *PouchDaemonSuite) TestDaemonDefaultRegistry(c *check.C) {
 	defer dcfg.KillDaemon()
 }
 
-// TestDaemonCriEnbaled tests enabling cri part in pouchd.
-func (suite *PouchDaemonSuite) TestDaemonCriEnbaled(c *check.C) {
+// TestDaemonCriEnabled tests enabling cri part in pouchd.
+func (suite *PouchDaemonSuite) TestDaemonCriEnabled(c *check.C) {
 	dcfg, err := StartDefaultDaemonDebug(
 		"--enable-cri")
 	c.Assert(err, check.IsNil)
 
 	result := RunWithSpecifiedDaemon(dcfg, "info")
-	err = util.PartialEqual(result.Combined(), "CriEnabled: true")
+	err = util.PartialEqual(result.Combined(), "CriEnabled:  true")
 	c.Assert(err, check.IsNil)
 
 	defer dcfg.KillDaemon()
